Meal prep rendering.
Meal prep gets rendered two slightly different ways. For week
templates, we show non-prepped meals (like eating out and
ones that must be cooked-to-order) to get a fuller take at a
glance of what that week will be like. (This is more like a "week summary" than a meal prep.

But for the meal prep  section itself, we only render those that
meals that do require some kind of prep. This is so we can add
those to some kind of todo list when prepping for the week.